+++ src/app/universes/universe-characters-resolver.service.ts
@@ -5,7 +5,9 @@ import { map, switchMap, tap } from "rxjs/operators";
 
 import { CharacterReference } from "../characters/characters.model";
 import { CharacterStateService } from "../characters/shared/character-state.service";
-import { CharacterQueryResult, CharacterService } from "../core/character.service";
+import { CharacterQuery, CharacterQueryResult, CharacterService } from "../core/character.service";
+
+import { Universe } from "./shared/universe.model";
 
 @Injectable({
     providedIn: "root",
@@ -15,9 +17,17 @@ export class UniverseCharactersResolverService implements Resolve<Observable<Cha
 
     public resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
         const universeId = route.params["universeId"];
+        const universe = route.parent.data["universe"] as Universe;
+
+        const query = {
+            sort: universe.settings.allowLexicographicalOrdering ? "lexicographical" : "nominal",
+            includeHidden: true,
+            page: 0,
+            query: "",
+        } as CharacterQuery;
         this.characterStateService.clearCharactersAndReferences();
         return this.characterService
-            .getCharacters(universeId)
+            .getCharacters(universeId, query)
             .pipe(tap((r) => this.characterStateService.addReferences(...r.characters)));
     }
 }