+++ src/environments/environment.prod.ts
@@ -1,6 +1,6 @@
 export const environment = {
     production: true,
     characterPageLimit: 20,
-    version: "0.4.2",
+    version: "0.4.4",
     apiEndpoint: "https://api.characterbase.xyz",
 };