+++ src/environments/environment.ts
@@ -5,7 +5,7 @@
 export const environment = {
     production: false,
     characterPageLimit: 5,
-    version: "0.4.3",
+    version: "0.4.4",
     apiEndpoint: "http://localhost:3000",
 };